frank.py: Add fallback for code size mismatch
If the profile code size does not match, the output will almost
certainly be invalid, and with the previous .text-section-only patch
this will fail an assertion prior to file writing.

However, this would be confusing to anyone trying to match code who
isn't familiar with frank - so let's add a fallback here to output a
warning instead, so they can see the asmdiff mismatch and proceed with
`ifndef MUST_MATCH` guarding as usual.